+++ mtest/mpi-types.h
@@ -5,9 +5,9 @@ typedef unsigned int       mp_word;   /* 4 byte type */
 typedef unsigned int       mp_size;
 typedef int                mp_err;
 
-#define MP_DIGIT_BIT       (CHAR_BIT*sizeof(mp_digit))
+#define MP_DIGIT_BIT       (CHAR_BIT*(int)sizeof(mp_digit))
 #define MP_DIGIT_MAX       USHRT_MAX
-#define MP_WORD_BIT        (CHAR_BIT*sizeof(mp_word))
+#define MP_WORD_BIT        (CHAR_BIT*(int)sizeof(mp_word))
 #define MP_WORD_MAX        UINT_MAX
 
 #define MP_DIGIT_SIZE      2
